requirements_phase2_minimal.txt
langdetect==1.0.9
deep-translator==1.11.4
translate==3.6.1
textblob==0.17.1
python-dotenv==1.0.0
pydantic==2.5.3
pytest==7.4.4
requests==2.31.0
tqdm==4.66.1
click==8.1.7
PyYAML==6.0.1

# Phase 3: Web Interface dependencies
fastapi==0.115.7
uvicorn==0.32.1
python-multipart==0.0.18
jinja2==3.1.4